requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
awscli==1.14.58
backports.ssl-match-hostname==3.5.0.1
boto3==1.6.11
botocore==1.9.11
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.7
constantly==15.1.0
cryptography==2.1.4
cssselect==1.0.3
docker-py==1.10.6
docker-pycreds==0.2.2
docopt==0.6.2
docutils==0.14
enum34==1.1.6
futures==3.2.0
gitignore==0.0.8
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
ipaddress==1.0.19
jmespath==0.9.3
lxml==4.2.0
parsel==1.4.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
pymongo==3.6.1
pyOpenSSL==17.5.0
python-dateutil==2.1
PyYAML==3.12
queuelib==1.5.0
requests==2.18.4
retrying==1.3.3
rsa==3.4.2
s3transfer==0.1.13
scrapinghub==2.0.3
Scrapy==1.5.0
service-identity==17.0.0
shub==2.7.0
six==1.11.0
tqdm==4.19.7
Twisted==17.9.0
urllib3==1.22
w3lib==1.19.0
websocket-client==0.47.0
zope.interface==4.4.3
